# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crafted By API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a list of users",
                "description": "Display a listing of the resource.",
                "operationId": "5a61fd217e7500976e1ae5c5662a4903",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserCard"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Users Store",
                "description": "Create User",
                "operationId": "addUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserStore"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserCard"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/users/{user_id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Find a user by Id",
                "description": "Return a single user",
                "operationId": "getUserById",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "Id of user to return",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update existing user",
                "description": "Returns updated user data",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserModel"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete existing user",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Login User",
                "description": "Login User",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "example": "admin@admin.admin"
                                    },
                                    "password": {
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "accessToken": {
                                            "description": "Bearer token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "description": "Token type",
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserCard"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BusinessCard": {
                "title": "Business Card",
                "description": "Small Business Data",
                "properties": {
                    "Id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "uuid"
                    },
                    "name": {
                        "title": "Business name",
                        "type": "string"
                    },
                    "email": {
                        "title": "Business email",
                        "type": "string"
                    },
                    "theme": {
                        "$ref": "#/components/schemas/Theme"
                    }
                },
                "type": "object"
            },
...
            "Theme": {
                "title": "Business Theme",
                "properties": {
                    "layer": {
                        "title": "Theme Layer",
                        "type": "string"
                    },
                    "color1": {
                        "title": "Theme color n°1",
                        "type": "string"
                    },
                    "color2": {
                        "title": "Theme color n°2",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserCard": {
                "title": "User card",
                "description": "Small User data",
                "properties": {
                    "Id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "uuid"
                    },
                    "lastname": {
                        "title": "User Lastname",
                        "type": "string"
                    },
                    "firstname": {
                        "title": "User firstname",
                        "type": "string"
                    },
                    "role": {
                        "title": "User role",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserStore": {
                "title": "User Store",
                "description": "User model to store",
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "Id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "uuid"
                    },
                    "lastname": {
                        "title": "Lastname",
                        "type": "string",
                        "example": "Hureau"
                    },
                    "firstname": {
                        "title": "Firstname",
                        "type": "string",
                        "example": "Nicolas"
                    },
                    "email": {
                        "title": "Email",
                        "type": "email",
                        "example": "nicolas.hureau@crafted-by.com"
                    },
                    "password": {
                        "title": "password",
                        "required": [
                            "email",
                            "password"
                        ],
                        "type": "password",
                        "example": "password"
                    },
                    "password_confirmation": {
                        "type": "password",
                        "example": "password"
                    },
                    "address": {
                        "title": "Address",
                        "type": "string",
                        "example": "Papeteries"
                    },
                    "zip_code": {
                        "title": "Postal Code",
                        "type": "string",
                        "example": "74960"
                    },
                    "city": {
                        "title": "City",
                        "type": "string",
                        "example": "Cran-Gevrier"
                    }
                },
                "type": "object"
            },
            "UserModel": {
                "title": "UserModel",
                "description": "User Model",
                "properties": {
                    "Id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "uuid"
                    },
                    "lastname": {
                        "title": "Lastname",
                        "type": "string",
                        "example": "Hureau"
                    },
                    "firstname": {
                        "title": "Firstname",
                        "type": "string",
                        "example": "Nicolas"
                    },
                    "email": {
                        "title": "Email",
                        "type": "email",
                        "example": "nicolas.hureau@crafted-by.com"
                    },
                    "address": {
                        "title": "Address",
                        "type": "string",
                        "example": "Papeteries"
                    },
                    "zip_code": {
                        "title": "Postal Code",
                        "type": "string",
                        "example": "74960"
                    },
                    "city": {
                        "title": "City",
                        "type": "string",
                        "example": "Cran-Gevrier"
                    },
                    "business": {
                        "$ref": "#/components/schemas/BusinessCard"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}